Clean up Gun file
