+++ Assets/Code/Scripts/AmmoCount.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fcda1f3bb56990d46bf1f896bc38d1ef
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
